requirements.txt
celery==5.2.7
coverage==7.2.5
cerberus==1.3.4


django>=4.2.0,<4.3
djangorestframework==3.14.0
django-filter==23.2
djangorestframework_simplejwt==5.2.2
django-dotenv==1.4.2
django-jazzmin>=2.6.0,<2.7
drf-yasg==1.21.5

flower==1.2.0
flake8==6.0.0

model_bakery==1.11.0

psycopg2-binary==2.9.6
Pillow==9.5.0

redis==4.5.4

uWSGI>=2.0.19.1,<2.1


